Clarify signal dependency between multiprocess classes
Using constants defined in signals as an attempt to make shared protocol
between various generic and implemented classes more explicit.
GenericVHF main_func has absorbed the definition from the prior
run_vhf_dep/collect.py in the process, as the exact signal in the
main_func loop should not be a specific concern to the implementing
class.
